[CP-2901] `DeviceState` moved to `device-manager` domain
