+++ libs/core-device/feature/src/selectors/get-configured-core-devices.selector.ts
@@ -4,7 +4,8 @@
  */
 
 import { createSelector } from "@reduxjs/toolkit"
-import { Device, DeviceState } from "core-device/models"
+import { Device } from "core-device/models"
+import { DeviceState } from "device-manager/models"
 import { getCoreDevicesSelector } from "./get-core-devices.selector"
 
 export const getConfiguredCoreDevicesSelector = createSelector(
