+++ libs/core-device/models/src/core-device.interface.ts
@@ -8,19 +8,11 @@ import { DeviceType } from "device-protocol/models"
 import { DeviceProperties } from "device-manager/models"
 import { DeviceId } from "Core/device/constants/device-id"
 
-export enum DeviceState {
-  Connected = "CONNECTED",
-  Configured = "CONFIGURED",
-  Initialized = "INITIALIZED",
-  Failed = "FAILED",
-}
-
 export interface Device extends DeviceProperties {
   id: DeviceId
   serialNumber: string | undefined
   deviceType: DeviceType
   caseColour: CaseColour | undefined
-  state: DeviceState
 }
 
 export interface CoreDeviceState {
