+++ libs/device-manager/models/src/device-properties.interface.ts
@@ -6,6 +6,14 @@
 import { DeviceBaseProperties } from "device-protocol/models"
 import { CaseColour } from "core-device/models"
 
+export enum DeviceState {
+  Connected = "CONNECTED",
+  Configured = "CONFIGURED",
+  Initialized = "INITIALIZED",
+  Failed = "FAILED",
+}
+
 export interface DeviceProperties extends DeviceBaseProperties {
   caseColour?: CaseColour
+  state: DeviceState
 }
