+++ libs/generic-view/store/src/lib/hooks/use-api-serial-port-listeners.ts
@@ -7,7 +7,7 @@ import { useCallback, useEffect } from "react"
 import { useDispatch, useSelector } from "react-redux"
 import { answerMain, useDebouncedEventsHandler } from "shared/utils"
 import { DeviceProtocolMainEvent, DeviceType } from "device-protocol/models"
-import { DeviceState } from "generic-view/models"
+import { DeviceState } from "device-manager/models"
 import { DeviceBaseProperties } from "device-protocol/models"
 import { Dispatch } from "Core/__deprecated__/renderer/store"
 import { addDevice, removeDevice } from "../views/actions"
