+++ libs/generic-view/store/src/lib/views/reducer.ts
@@ -6,7 +6,8 @@
 import { createReducer } from "@reduxjs/toolkit"
 import { MenuElement } from "Core/__deprecated__/renderer/constants/menu-elements"
 import { View } from "generic-view/utils"
-import { Device, DeviceState, Features } from "generic-view/models"
+import { DeviceState } from "device-manager/models"
+import { Device, Features } from "generic-view/models"
 import { ApiError } from "device/models"
 import { AppError } from "Core/core/errors"
 import { getAPIConfig } from "../get-api-config"
